+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Block(`@amsterdam/navigation`): Add navigation block.
 - Editor: Add support for deleting apps.
 - Server: Add support for deleting apps.