+++ CHANGELOG.md
@@ -7,6 +7,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.8.8] - 2019-10-01
+
 ### Added
 
 - Block(`@appsemble/form`): Add support for field icons.